othello: project othello-build-tools added
This commit adds the othello-build-tools project which contains the
configuration files for various static analysis plug-ins operating on Java
source code and Java byte code.

Note that this project and the configuration of associated Maven plug-ins have
been deliberately kept separated from (and redundant to) the gaming-build-tools
project in order to be able to build the Othello projects without checking out
the gaming repository, as POM inheritance requires the parent POM to be
immediately available. In addition, this allows independent evolution of the
plug-ins' versions and configurations without breaking things.