othello: project othello-core added
This commit adds an implementation of the Othello game. It provides the game
elements as well as the two types of moves (placing a token / skipping a move).
Individual strategies for choosing the next move are to be provided by separate
projects using Java's SPI mechanism.

Note that this project does not come with a corresponding game visualisation.
This will be provided by a separate project.